source/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.2
Django==4.1.7
django-bootstrap-v5==1.0.11
djangorestframework==3.14.0
psycopg2-binary==2.9.5
pytz==2023.3
soupsieve==2.4
sqlparse==0.4.3
